backend/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
mysql-connector-python==8.0.28
librosa==0.9.1
pydub==0.25.1
requests==2.26.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.3.24
mysqlclient==2.1.0
beautifulsoup4==4.10.0  # or the latest version
requests==2.26.0